@@ -1,3 +1,6 @@
+// Copyright © 2024 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 package opts
 
 import (
@@ -9,7 +12,6 @@ import (
 
 	"github.com/docker/cli/cli/compose/loader"
 	units "github.com/docker/go-units"
-	"github.com/pkg/errors"
 )
 
 var (
@@ -342,7 +344,7 @@ func (m *MemBytes) UnmarshalJSON(s []byte) error {
 func MountParser(mount string) (source, destination string, err error) {
 	spec, err := loader.ParseVolume(mount)
 	if err != nil {
-		return "", "", errors.Wrap(err, "Failed to parse mount")
+		return "", "", fmt.Errorf("Failed to parse mount: %w", err)
 	}
 
 	return spec.Source, spec.Target, nil
