@@ -1,3 +1,6 @@
+// Copyright © 2024 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 // Copyright 2018 go-dockerclient authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
@@ -7,7 +10,7 @@ package docker
 import (
 	"context"
 	"encoding/json"
-	"io/ioutil"
+	"io"
 	"net/http"
 )
 
@@ -40,10 +43,10 @@ func (c *Client) InstallPlugins(opts InstallPluginOptions) error {
 		data:    opts.Plugins,
 		context: opts.Context,
 	})
-	defer resp.Body.Close()
 	if err != nil {
 		return err
 	}
+	defer resp.Body.Close()
 	return nil
 }
 
@@ -288,11 +291,10 @@ type EnablePluginOptions struct {
 func (c *Client) EnablePlugin(opts EnablePluginOptions) error {
 	path := "/plugins/" + opts.Name + "/enable?" + queryString(opts)
 	resp, err := c.do("POST", path, doOptions{context: opts.Context})
-	defer resp.Body.Close()
 	if err != nil {
 		return err
 	}
-	resp.Body.Close()
+	defer resp.Body.Close()
 	return nil
 }
 
@@ -312,11 +314,10 @@ type DisablePluginOptions struct {
 func (c *Client) DisablePlugin(opts DisablePluginOptions) error {
 	path := "/plugins/" + opts.Name + "/disable"
 	resp, err := c.do("POST", path, doOptions{context: opts.Context})
-	defer resp.Body.Close()
 	if err != nil {
 		return err
 	}
-	resp.Body.Close()
+	defer resp.Body.Close()
 	return nil
 }
 
@@ -340,11 +341,11 @@ func (c *Client) CreatePlugin(opts CreatePluginOptions) (string, error) {
 	resp, err := c.do("POST", path, doOptions{
 		data:    opts.Path,
 		context: opts.Context})
-	defer resp.Body.Close()
 	if err != nil {
 		return "", err
 	}
-	containerNameBytes, err := ioutil.ReadAll(resp.Body)
+	defer resp.Body.Close()
+	containerNameBytes, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return "", err
 	}
@@ -367,10 +368,10 @@ type PushPluginOptions struct {
 func (c *Client) PushPlugin(opts PushPluginOptions) error {
 	path := "/plugins/" + opts.Name + "/push"
 	resp, err := c.do("POST", path, doOptions{context: opts.Context})
-	defer resp.Body.Close()
 	if err != nil {
 		return err
 	}
+	defer resp.Body.Close()
 	return nil
 }
 
@@ -394,13 +395,13 @@ func (c *Client) ConfigurePlugin(opts ConfigurePluginOptions) error {
 		data:    opts.Envs,
 		context: opts.Context,
 	})
-	defer resp.Body.Close()
 	if err != nil {
 		if e, ok := err.(*Error); ok && e.Status == http.StatusNotFound {
 			return &NoSuchPlugin{ID: opts.Name}
 		}
 		return err
 	}
+	defer resp.Body.Close()
 	return nil
 }
 
