@@ -1,9 +1,12 @@
+// Copyright © 2024 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 package dockertest
 
 import (
+	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"os"
 	"path/filepath"
@@ -14,7 +17,6 @@ import (
 	"github.com/cenkalti/backoff/v4"
 	dc "github.com/ory/dockertest/v3/docker"
 	options "github.com/ory/dockertest/v3/docker/opts"
-	"github.com/pkg/errors"
 )
 
 var (
@@ -69,7 +71,11 @@ func (r *Resource) GetBoundIP(id string) string {
 		return ""
 	}
 
-	return m[0].HostIP
+	ip := m[0].HostIP
+	if ip == "0.0.0.0" || ip == "" {
+		return "localhost"
+	}
+	return ip
 }
 
 // GetHostPort returns a resource's published port with an address.
@@ -84,7 +90,7 @@ func (r *Resource) GetHostPort(portID string) string {
 	}
 
 	ip := m[0].HostIP
-	if ip == "0.0.0.0" {
+	if ip == "0.0.0.0" || ip == "" {
 		ip = "localhost"
 	}
 	return net.JoinHostPort(ip, m[0].HostPort)
@@ -113,13 +119,23 @@ func (r *Resource) Exec(cmd []string, opts ExecOptions) (exitCode int, err error
 		Container:    r.Container.ID,
 		Cmd:          cmd,
 		Env:          opts.Env,
-		AttachStderr: opts.StdErr != nil,
-		AttachStdout: opts.StdOut != nil,
+		AttachStderr: true,
+		AttachStdout: true,
 		AttachStdin:  opts.StdIn != nil,
 		Tty:          opts.TTY,
 	})
 	if err != nil {
-		return -1, errors.Wrap(err, "Create exec failed")
+		return -1, fmt.Errorf("Create exec failed: %w", err)
+	}
+
+	// Always attach stderr/stdout, even if not specified, to ensure that exec
+	// waits with opts.Detach as false (default)
+	// ref: https://github.com/fsouza/go-dockerclient/issues/838
+	if opts.StdErr == nil {
+		opts.StdErr = io.Discard
+	}
+	if opts.StdOut == nil {
+		opts.StdOut = io.Discard
 	}
 
 	err = r.pool.Client.StartExec(exec.ID, dc.StartExecOptions{
@@ -129,12 +145,12 @@ func (r *Resource) Exec(cmd []string, opts ExecOptions) (exitCode int, err error
 		Tty:          opts.TTY,
 	})
 	if err != nil {
-		return -1, errors.Wrap(err, "Start exec failed")
+		return -1, fmt.Errorf("Start exec failed: %w", err)
 	}
 
 	inspectExec, err := r.pool.Client.InspectExec(exec.ID)
 	if err != nil {
-		return -1, errors.Wrap(err, "Inspect exec failed")
+		return -1, fmt.Errorf("Inspect exec failed: %w", err)
 	}
 
 	return inspectExec.ExitCode, nil
@@ -161,18 +177,18 @@ func (r *Resource) ConnectToNetwork(network *Network) error {
 		dc.NetworkConnectionOptions{Container: r.Container.ID},
 	)
 	if err != nil {
-		return errors.Wrap(err, "Failed to connect container to network")
+		return fmt.Errorf("Failed to connect container to network: %w", err)
 	}
 
 	// refresh internal representation
 	r.Container, err = r.pool.Client.InspectContainer(r.Container.ID)
 	if err != nil {
-		return errors.Wrap(err, "Failed to refresh container information")
+		return fmt.Errorf("Failed to refresh container information: %w", err)
 	}
 
 	network.Network, err = r.pool.Client.NetworkInfo(network.Network.ID)
 	if err != nil {
-		return errors.Wrap(err, "Failed to refresh network information")
+		return fmt.Errorf("Failed to refresh network information: %w", err)
 	}
 
 	return nil
@@ -185,18 +201,18 @@ func (r *Resource) DisconnectFromNetwork(network *Network) error {
 		dc.NetworkConnectionOptions{Container: r.Container.ID},
 	)
 	if err != nil {
-		return errors.Wrap(err, "Failed to connect container to network")
+		return fmt.Errorf("Failed to connect container to network: %w", err)
 	}
 
 	// refresh internal representation
 	r.Container, err = r.pool.Client.InspectContainer(r.Container.ID)
 	if err != nil {
-		return errors.Wrap(err, "Failed to refresh container information")
+		return fmt.Errorf("Failed to refresh container information: %w", err)
 	}
 
 	network.Network, err = r.pool.Client.NetworkInfo(network.Network.ID)
 	if err != nil {
-		return errors.Wrap(err, "Failed to refresh network information")
+		return fmt.Errorf("Failed to refresh network information: %w", err)
 	}
 
 	return nil
@@ -226,7 +242,7 @@ func NewTLSPool(endpoint, certpath string) (*Pool, error) {
 
 	client, err := dc.NewTLSClient(endpoint, cert, key, ca)
 	if err != nil {
-		return nil, errors.Wrap(err, "")
+		return nil, err
 	}
 
 	return &Pool{
@@ -243,7 +259,7 @@ func NewPool(endpoint string) (*Pool, error) {
 		if os.Getenv("DOCKER_MACHINE_NAME") != "" {
 			client, err := dc.NewClientFromEnv()
 			if err != nil {
-				return nil, errors.Wrap(err, "failed to create client from environment")
+				return nil, fmt.Errorf("failed to create client from environment: %w", err)
 			}
 
 			return &Pool{Client: client}, nil
@@ -253,9 +269,13 @@ func NewPool(endpoint string) (*Pool, error) {
 		} else if os.Getenv("DOCKER_URL") != "" {
 			endpoint = os.Getenv("DOCKER_URL")
 		} else if runtime.GOOS == "windows" {
-			endpoint = "http://localhost:2375"
+			if _, err := os.Stat(`\\.\pipe\docker_engine`); err == nil {
+				endpoint = "npipe:////./pipe/docker_engine"
+			} else {
+				endpoint = "http://localhost:2375"
+			}
 		} else {
-			endpoint = "unix:///var/run/docker.sock"
+			endpoint = options.DefaultHost
 		}
 	}
 
@@ -265,7 +285,7 @@ func NewPool(endpoint string) (*Pool, error) {
 
 	client, err := dc.NewClient(endpoint)
 	if err != nil {
-		return nil, errors.Wrap(err, "")
+		return nil, err
 	}
 
 	return &Pool{
@@ -302,6 +322,7 @@ type RunOptions struct {
 	Privileged   bool
 	User         string
 	Tty          bool
+	Platform     string
 }
 
 // BuildOptions is used to pass in optional parameters when building a container
@@ -310,6 +331,9 @@ type BuildOptions struct {
 	ContextDir string
 	BuildArgs  []dc.BuildArg
 	Platform   string
+	// Version specifies the builder to use. "1" for classic, "2" for BuildKit
+	Version string
+	Auth    dc.AuthConfigurations
 }
 
 // BuildAndRunWithBuildOptions builds and starts a docker container.
@@ -318,14 +342,16 @@ func (d *Pool) BuildAndRunWithBuildOptions(buildOpts *BuildOptions, runOpts *Run
 	err := d.Client.BuildImage(dc.BuildImageOptions{
 		Name:         runOpts.Name,
 		Dockerfile:   buildOpts.Dockerfile,
-		OutputStream: ioutil.Discard,
+		OutputStream: io.Discard,
 		ContextDir:   buildOpts.ContextDir,
 		BuildArgs:    buildOpts.BuildArgs,
 		Platform:     buildOpts.Platform,
+		Version:      buildOpts.Version,
+		AuthConfigs:  buildOpts.Auth,
 	})
 
 	if err != nil {
-		return nil, errors.Wrap(err, "")
+		return nil, err
 	}
 
 	runOpts.Repository = runOpts.Name
@@ -350,10 +376,10 @@ func (d *Pool) BuildAndRun(name, dockerfilePath string, env []string) (*Resource
 // RunWithOptions starts a docker container.
 // Optional modifier functions can be passed in order to change the hostconfig values not covered in RunOptions
 //
-//  pool.RunWithOptions(&RunOptions{Repository: "mongo", Cmd: []string{"mongod", "--smallfiles"}})
-//  pool.RunWithOptions(&RunOptions{Repository: "mongo", Cmd: []string{"mongod", "--smallfiles"}}, func(hostConfig *dc.HostConfig) {
-//			hostConfig.ShmSize = shmemsize
-//		})
+//	 pool.RunWithOptions(&RunOptions{Repository: "mongo", Cmd: []string{"mongod", "--smallfiles"}})
+//	 pool.RunWithOptions(&RunOptions{Repository: "mongo", Cmd: []string{"mongod", "--smallfiles"}}, func(hostConfig *dc.HostConfig) {
+//				hostConfig.ShmSize = shmemsize
+//			})
 func (d *Pool) RunWithOptions(opts *RunOptions, hcOpts ...func(*dc.HostConfig)) (*Resource, error) {
 	repository := opts.Repository
 	tag := opts.Tag
@@ -400,11 +426,24 @@ func (d *Pool) RunWithOptions(opts *RunOptions, hcOpts ...func(*dc.HostConfig)) 
 
 	_, err := d.Client.InspectImage(fmt.Sprintf("%s:%s", repository, tag))
 	if err != nil {
+		var (
+			auth  = opts.Auth
+			parts = strings.SplitN(repository, "/", 3)
+			empty = opts.Auth == dc.AuthConfiguration{}
+		)
+		if empty && len(parts) == 3 {
+			res, err := dc.NewAuthConfigurationsFromCredsHelpers(parts[0])
+			if err == nil {
+				auth = *res
+			}
+		}
+
 		if err := d.Client.PullImage(dc.PullImageOptions{
 			Repository: repository,
 			Tag:        tag,
-		}, opts.Auth); err != nil {
-			return nil, errors.Wrap(err, "")
+			Platform:   opts.Platform,
+		}, auth); err != nil {
+			return nil, err
 		}
 	}
 
@@ -444,22 +483,22 @@ func (d *Pool) RunWithOptions(opts *RunOptions, hcOpts ...func(*dc.HostConfig)) 
 		NetworkingConfig: &networkingConfig,
 	})
 	if err != nil {
-		return nil, errors.Wrap(err, "")
+		return nil, err
 	}
 
 	if err := d.Client.StartContainer(c.ID, nil); err != nil {
-		return nil, errors.Wrap(err, "")
+		return nil, err
 	}
 
-	c, err = d.Client.InspectContainer(c.ID)
+	c, err = d.inspectContainerWithRetries(c.ID)
 	if err != nil {
-		return nil, errors.Wrap(err, "")
+		return nil, err
 	}
 
 	for _, network := range opts.Networks {
 		network.Network, err = d.Client.NetworkInfo(network.Network.ID)
 		if err != nil {
-			return nil, errors.Wrap(err, "")
+			return nil, err
 		}
 	}
 
@@ -469,9 +508,40 @@ func (d *Pool) RunWithOptions(opts *RunOptions, hcOpts ...func(*dc.HostConfig)) 
 	}, nil
 }
 
+// inspectContainerWithRetries will repeat the inspect call until the container has port bindings assigned.
+func (d *Pool) inspectContainerWithRetries(id string) (*dc.Container, error) {
+	const maxRetries = 10
+	var (
+		retryNum int
+		c        *dc.Container
+		err      error
+	)
+	for retryNum <= maxRetries {
+		if retryNum > 0 {
+			time.Sleep(100 * time.Millisecond)
+		}
+		c, err = d.Client.InspectContainer(id)
+		if err != nil {
+			return nil, err
+		}
+		if hasEmptyPortBindings := func() bool {
+			for _, bindings := range c.NetworkSettings.Ports {
+				if len(bindings) == 0 {
+					return true
+				}
+			}
+			return false
+		}(); !hasEmptyPortBindings {
+			return c, nil
+		}
+		retryNum++
+	}
+	return c, err
+}
+
 // Run starts a docker container.
 //
-//  pool.Run("mysql", "5.3", []string{"FOO=BAR", "BAR=BAZ"})
+//	pool.Run("mysql", "5.3", []string{"FOO=BAR", "BAR=BAZ"})
 func (d *Pool) Run(repository, tag string, env []string) (*Resource, error) {
 	return d.RunWithOptions(&RunOptions{Repository: repository, Tag: tag, Env: env})
 }
@@ -509,11 +579,11 @@ func (d *Pool) RemoveContainerByName(containerName string) error {
 	containers, err := d.Client.ListContainers(dc.ListContainersOptions{
 		All: true,
 		Filters: map[string][]string{
-			"name": []string{containerName},
+			"name": {containerName},
 		},
 	})
 	if err != nil {
-		return errors.Wrapf(err, "Error while listing containers with name %s", containerName)
+		return fmt.Errorf("Error while listing containers with name %s: %w", containerName, err)
 	}
 
 	if len(containers) == 0 {
@@ -526,7 +596,7 @@ func (d *Pool) RemoveContainerByName(containerName string) error {
 		RemoveVolumes: true,
 	})
 	if err != nil {
-		return errors.Wrapf(err, "Error while removing container with name %s", containerName)
+		return fmt.Errorf("Error while removing container with name %s: %w", containerName, err)
 	}
 
 	return nil
@@ -535,7 +605,7 @@ func (d *Pool) RemoveContainerByName(containerName string) error {
 // Purge removes a container and linked volumes from docker.
 func (d *Pool) Purge(r *Resource) error {
 	if err := d.Client.RemoveContainer(dc.RemoveContainerOptions{ID: r.Container.ID, Force: true, RemoveVolumes: true}); err != nil {
-		return errors.Wrap(err, "")
+		return err
 	}
 
 	return nil
@@ -551,7 +621,7 @@ func (d *Pool) Retry(op func() error) error {
 	bo.MaxElapsedTime = d.MaxWait
 	if err := backoff.Retry(op, bo); err != nil {
 		if bo.NextBackOff() == backoff.Stop {
-			return fmt.Errorf("reached retry deadline")
+			return fmt.Errorf("reached retry deadline: %w", err)
 		}
 
 		return err
@@ -566,7 +636,7 @@ func (d *Pool) CurrentContainer() (*Resource, error) {
 	// docker daemon puts short container id into hostname
 	hostname, err := os.Hostname()
 	if err != nil {
-		return nil, errors.Wrap(err, "Get hostname failed")
+		return nil, fmt.Errorf("Get hostname failed: %w", err)
 	}
 
 	container, err := d.Client.InspectContainer(hostname)
@@ -579,7 +649,7 @@ func (d *Pool) CurrentContainer() (*Resource, error) {
 	case *dc.NoSuchContainer:
 		return nil, ErrNotInContainer
 	default:
-		return nil, errors.Wrap(err, "")
+		return nil, err
 	}
 }
 
@@ -593,13 +663,35 @@ func (d *Pool) CreateNetwork(name string, opts ...func(config *dc.CreateNetworkO
 
 	network, err := d.Client.CreateNetwork(cfg)
 	if err != nil {
-		return nil, errors.Wrap(err, "")
+		return nil, err
 	}
 
 	return &Network{
 		pool:    d,
 		Network: network,
 	}, nil
+}
+
+// NetworksByName returns a list of docker networks filtered by name
+func (d *Pool) NetworksByName(name string) ([]Network, error) {
+	networks, err := d.Client.ListNetworks()
+	if err != nil {
+		return nil, err
+	}
+
+	var foundNetworks []Network
+	for idx := range networks {
+		if networks[idx].Name == name {
+			foundNetworks = append(foundNetworks,
+				Network{
+					pool:    d,
+					Network: &networks[idx],
+				},
+			)
+		}
+	}
+
+	return foundNetworks, nil
 }
 
 // RemoveNetwork disconnects containers and removes provided network.
