@@ -1,3 +1,6 @@
+// Copyright © 2024 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 package opts
 
 import (
@@ -5,8 +8,6 @@ import (
 	"os"
 	"runtime"
 	"strings"
-
-	"github.com/pkg/errors"
 )
 
 // ValidateEnv validates an environment variable and returns it.
@@ -20,7 +21,7 @@ import (
 func ValidateEnv(val string) (string, error) {
 	arr := strings.Split(val, "=")
 	if arr[0] == "" {
-		return "", errors.Errorf("invalid environment variable: %s", val)
+		return "", fmt.Errorf("invalid environment variable: %s", val)
 	}
 	if len(arr) > 1 {
 		return val, nil
