@@ -1,3 +1,6 @@
+// Copyright © 2024 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 package types // import "github.com/ory/dockertest/v3/docker/types"
 
 import (
@@ -5,9 +8,9 @@ import (
 	"io"
 	"net"
 
+	units "github.com/docker/go-units"
 	"github.com/ory/dockertest/v3/docker/types/container"
 	"github.com/ory/dockertest/v3/docker/types/filters"
-	units "github.com/docker/go-units"
 )
 
 // CheckpointCreateOptions holds parameters to create a checkpoint from a container
@@ -240,7 +243,7 @@ type ImagePullOptions struct {
 // if the privilege request fails.
 type RequestPrivilegeFunc func() (string, error)
 
-//ImagePushOptions holds information to push images.
+// ImagePushOptions holds information to push images.
 type ImagePushOptions ImagePullOptions
 
 // ImageRemoveOptions holds parameters to remove images.
