@@ -1,10 +1,13 @@
+// Copyright © 2024 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 package container // import "github.com/ory/dockertest/v3/docker/types/container"
 
 import (
 	"time"
 
-	"github.com/ory/dockertest/v3/docker/types/strslice"
 	"github.com/docker/go-connections/nat"
+	"github.com/ory/dockertest/v3/docker/types/strslice"
 )
 
 // MinimumDuration puts a minimum on user configured duration.
