@@ -1,13 +1,16 @@
+// Copyright © 2024 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 package container // import "github.com/ory/dockertest/v3/docker/types/container"
 
 import (
 	"strings"
 
+	"github.com/docker/go-connections/nat"
+	"github.com/docker/go-units"
 	"github.com/ory/dockertest/v3/docker/types/blkiodev"
 	"github.com/ory/dockertest/v3/docker/types/mount"
 	"github.com/ory/dockertest/v3/docker/types/strslice"
-	"github.com/docker/go-connections/nat"
-	"github.com/docker/go-units"
 )
 
 // Isolation represents the isolation technology of a container. The supported
@@ -122,7 +125,7 @@ func (n NetworkMode) ConnectedContainer() string {
 	return ""
 }
 
-//UserDefined indicates user-created network
+// UserDefined indicates user-created network
 func (n NetworkMode) UserDefined() string {
 	if n.IsUserDefined() {
 		return string(n)
