@@ -1,11 +1,14 @@
+// Copyright © 2024 Ory Corp
+// SPDX-License-Identifier: Apache-2.0
+
 // Package v1p19 provides specific API types for the API version 1, patch 19.
 package v1p19 // import "github.com/ory/dockertest/v3/docker/types/versions/v1p19"
 
 import (
+	"github.com/docker/go-connections/nat"
 	"github.com/ory/dockertest/v3/docker/types"
 	"github.com/ory/dockertest/v3/docker/types/container"
 	"github.com/ory/dockertest/v3/docker/types/versions/v1p20"
-	"github.com/docker/go-connections/nat"
 )
 
 // ContainerJSON is a backcompatibility struct for APIs prior to 1.20.
